+++ samples/FluentlyHttpClient.Sample.Api/Program.cs
@@ -1,4 +1,6 @@
 
+using Microsoft.AspNetCore.Server.Kestrel.Core;
+
 namespace FluentlyHttpClient.Sample.Api;
 
 public class Program
@@ -10,7 +12,10 @@ public class Program
 			.ConfigureWebHostDefaults(webBuilder =>
 			{
 				webBuilder
-					.UseUrls("http://localhost:5500/")
+					.UseUrls("http://localhost:5500/", "https://localhost:5510")
+					.ConfigureKestrel(opts => opts
+						.ConfigureEndpointDefaults(lo => lo.Protocols = HttpProtocols.Http1AndHttp2AndHttp3)
+					)
 					.UseStartup<Startup>()
 					;
 			});